matrix: activate only the valid transformations on the semi-active matrix
In the previous implementation,
a matrix with hidden states will change between hidden states
of the inactive trait.
For example,
suppose that you have the traits 00a, 00b, 01, 10a, 10b, and 11
(state 0 of the second trait have two hidden states),
if the root is at 00 (either 00a or 00b)
and the first trait born,
then the only valid changes should be 00a <-> 10a,
and 00b <-> 10b.
In the previous implementation,
the changes 00a <-> 00b are accepted,
but this violates the assumption that the second trait is inactive.